requirements.txt
pymupdf==1.26.1
google-generativeai==0.8.3
opencv-python-headless==4.11.0.86
python-dotenv==1.0.1
pinecone-client==5.0.1
sentence_transformers==3.4.1
streamlit==1.42.2
tqdm==4.67.1

